[cli] refactor client functions into sub-modules
