+++ apps/src/bin/anoma-client/cli.rs
@@ -1,253 +1,67 @@
 //! Anoma client CLI.
 
-use std::collections::HashSet;
-use std::io::Write;
-
-use anoma::types::intent::{Exchange, FungibleTokenIntent};
-use anoma::types::key::ed25519::Signed;
-use anoma::types::{address, token};
 use anoma_apps::cli;
-use anoma_apps::cli::{args, cmds, Context};
-use anoma_apps::client::{rpc, signing, tx};
-use anoma_apps::config::genesis;
-use anoma_apps::node::ledger::tendermint_node;
-use anoma_apps::proto::services::rpc_service_client::RpcServiceClient;
-use anoma_apps::proto::{services, RpcMessage};
-use anoma_apps::wallet::Wallet;
-use borsh::BorshSerialize;
+use anoma_apps::cli::cmds::*;
+use anoma_apps::client::{gossip, rpc, tx, utils};
 use color_eyre::eyre::Result;
 
 pub async fn main() -> Result<()> {
     let (cmd, ctx) = cli::anoma_client_cli();
     match cmd {
         // Ledger cmds
-        cmds::AnomaClient::TxCustom(cmds::TxCustom(args)) => {
+        AnomaClient::TxCustom(TxCustom(args)) => {
             tx::submit_custom(ctx, args).await;
         }
-        cmds::AnomaClient::TxTransfer(cmds::TxTransfer(args)) => {
+        AnomaClient::TxTransfer(TxTransfer(args)) => {
             tx::submit_transfer(ctx, args).await;
         }
-        cmds::AnomaClient::TxUpdateVp(cmds::TxUpdateVp(args)) => {
+        AnomaClient::TxUpdateVp(TxUpdateVp(args)) => {
             tx::submit_update_vp(ctx, args).await;
         }
-        cmds::AnomaClient::TxInitAccount(cmds::TxInitAccount(args)) => {
+        AnomaClient::TxInitAccount(TxInitAccount(args)) => {
             tx::submit_init_account(ctx, args).await;
         }
-        cmds::AnomaClient::TxInitValidator(cmds::TxInitValidator(args)) => {
+        AnomaClient::TxInitValidator(TxInitValidator(args)) => {
             tx::submit_init_validator(ctx, args).await;
         }
-        cmds::AnomaClient::Bond(cmds::Bond(args)) => {
+        AnomaClient::Bond(Bond(args)) => {
             tx::submit_bond(ctx, args).await;
         }
-        cmds::AnomaClient::Unbond(cmds::Unbond(args)) => {
+        AnomaClient::Unbond(Unbond(args)) => {
             tx::submit_unbond(ctx, args).await;
         }
-        cmds::AnomaClient::Withdraw(cmds::Withdraw(args)) => {
+        AnomaClient::Withdraw(Withdraw(args)) => {
             tx::submit_withdraw(ctx, args).await;
         }
-        cmds::AnomaClient::QueryEpoch(cmds::QueryEpoch(args)) => {
-            rpc::query_epoch(args).await;
+        // Ledger queries
+        AnomaClient::QueryEpoch(QueryEpoch(args)) => {
+            rpc::query_epoch(ctx, args).await;
         }
-        cmds::AnomaClient::QueryBalance(cmds::QueryBalance(args)) => {
+        AnomaClient::QueryBalance(QueryBalance(args)) => {
             rpc::query_balance(ctx, args).await;
         }
-        cmds::AnomaClient::QueryBonds(cmds::QueryBonds(args)) => {
+        AnomaClient::QueryBonds(QueryBonds(args)) => {
             rpc::query_bonds(ctx, args).await;
         }
-        cmds::AnomaClient::QueryVotingPower(cmds::QueryVotingPower(args)) => {
+        AnomaClient::QueryVotingPower(QueryVotingPower(args)) => {
             rpc::query_voting_power(ctx, args).await;
         }
-        cmds::AnomaClient::QuerySlashes(cmds::QuerySlashes(args)) => {
+        AnomaClient::QuerySlashes(QuerySlashes(args)) => {
             rpc::query_slashes(ctx, args).await;
         }
         // Gossip cmds
-        cmds::AnomaClient::Intent(cmds::Intent(args)) => {
-            gossip_intent(ctx, args).await;
+        AnomaClient::Intent(Intent(args)) => {
+            gossip::gossip_intent(ctx, args).await;
         }
-        cmds::AnomaClient::SubscribeTopic(cmds::SubscribeTopic(args)) => {
-            subscribe_topic(ctx, args).await;
+        AnomaClient::SubscribeTopic(SubscribeTopic(args)) => {
+            gossip::subscribe_topic(ctx, args).await;
         }
         // Utils cmds
-        cmds::AnomaClient::Utils(cmd) => match cmd {
-            cmds::Utils::InitGenesisValidator(cmds::InitGenesisValidator(
-                args,
-            )) => init_genesis_validator(ctx, args),
+        AnomaClient::Utils(cmd) => match cmd {
+            Utils::InitGenesisValidator(InitGenesisValidator(args)) => {
+                utils::init_genesis_validator(ctx, args)
+            }
         },
     }
     Ok(())
-}
-
-async fn gossip_intent(
-    mut ctx: Context,
-    args::Intent {
-        node_addr,
-        topic,
-        source,
-        signing_key,
-        exchanges,
-        ledger_address,
-        to_stdout,
-    }: args::Intent,
-) {
-    let mut signed_exchanges: HashSet<Signed<Exchange>> =
-        HashSet::with_capacity(exchanges.len());
-    for exchange in exchanges {
-        let signed =
-            sign_exchange(exchange, &mut ctx.wallet, ledger_address.clone())
-                .await;
-        signed_exchanges.insert(signed);
-    }
-
-    let source_keypair = match ctx.get_opt_cached(signing_key) {
-        Some(key) => key,
-        None => {
-            let source = ctx.get_opt(source).unwrap_or_else(|| {
-                eprintln!("A source or a signing key is required.");
-                cli::safe_exit(1)
-            });
-            signing::find_keypair(
-                &mut ctx.wallet,
-                &source,
-                ledger_address.clone(),
-            )
-            .await
-        }
-    };
-    let signed_ft: Signed<FungibleTokenIntent> = Signed::new(
-        &source_keypair,
-        FungibleTokenIntent {
-            exchange: signed_exchanges,
-        },
-    );
-    let data_bytes = signed_ft.try_to_vec().unwrap();
-
-    if to_stdout {
-        let mut out = std::io::stdout();
-        out.write_all(&data_bytes).unwrap();
-        out.flush().unwrap();
-    } else {
-        let node_addr = node_addr.expect(
-            "Gossip node address must be defined to submit the intent to it.",
-        );
-        let topic = topic.expect(
-            "The topic must be defined to submit the intent to a gossip node.",
-        );
-        let mut client = RpcServiceClient::connect(node_addr).await.unwrap();
-
-        let intent = anoma::proto::Intent::new(data_bytes);
-        let message: services::RpcMessage =
-            RpcMessage::new_intent(intent, topic).into();
-        let response = client
-            .send_message(message)
-            .await
-            .expect("failed to send message and/or receive rpc response");
-        println!("{:#?}", response);
-    }
-}
-
-async fn sign_exchange(
-    exchange: Exchange,
-    wallet: &mut Wallet,
-    ledger_address: tendermint::net::Address,
-) -> Signed<Exchange> {
-    let source_keypair =
-        signing::find_keypair(wallet, &exchange.addr, ledger_address).await;
-    Signed::new(&source_keypair, exchange.clone())
-}
-
-async fn subscribe_topic(
-    _ctx: Context,
-    args::SubscribeTopic { node_addr, topic }: args::SubscribeTopic,
-) {
-    let mut client = RpcServiceClient::connect(node_addr).await.unwrap();
-    let message: services::RpcMessage = RpcMessage::new_topic(topic).into();
-    let response = client
-        .send_message(message)
-        .await
-        .expect("failed to send message and/or receive rpc response");
-    println!("{:#?}", response);
-}
-
-/// Initialize genesis validator's address, staking reward address, consensus
-/// key, validator account key and staking rewards key and use it in the
-/// ledger's node.",
-fn init_genesis_validator(
-    ctx: Context,
-    args::InitGenesisValidator {
-        alias,
-        unsafe_dont_encrypt,
-    }: args::InitGenesisValidator,
-) {
-    let mut wallet = ctx.wallet;
-
-    // Generate validator address
-    let validator_address =
-        address::gen_established_address("genesis validator address");
-    let validator_address_alias = alias.clone();
-    if !wallet
-        .add_address(validator_address_alias.clone(), validator_address.clone())
-    {
-        cli::safe_exit(1)
-    }
-    // Generate staking reward address
-    let rewards_address =
-        address::gen_established_address("genesis validator reward address");
-    let rewards_address_alias = format!("{}-rewards", alias);
-    if !wallet
-        .add_address(rewards_address_alias.clone(), rewards_address.clone())
-    {
-        cli::safe_exit(1)
-    }
-
-    println!("Generating validator account key...");
-    let (validator_key_alias, validator_key) = wallet.gen_key(
-        Some(format!("{}-validator-key", alias)),
-        unsafe_dont_encrypt,
-    );
-    println!("Generating consensus key...");
-    let (consensus_key_alias, consensus_key) = wallet.gen_key(
-        Some(format!("{}-consensus-key", alias)),
-        unsafe_dont_encrypt,
-    );
-    println!("Generating staking reward account key...");
-    let (rewards_key_alias, rewards_key) = wallet
-        .gen_key(Some(format!("{}-rewards-key", alias)), unsafe_dont_encrypt);
-
-    wallet.save().unwrap_or_else(|err| eprintln!("{}", err));
-
-    let tendermint_home = &ctx.config.ledger.tendermint;
-    tendermint_node::write_validator_key(
-        tendermint_home,
-        &validator_address,
-        &consensus_key,
-    );
-    tendermint_node::write_validator_state(tendermint_home);
-
-    println!();
-    println!("The validator's addresses and keys were stored in the wallet:");
-    println!("  Validator address \"{}\"", validator_address_alias);
-    println!("  Staking reward address \"{}\"", rewards_address_alias);
-    println!("  Validator account key \"{}\"", validator_key_alias);
-    println!("  Consensus key \"{}\"", consensus_key_alias);
-    println!("  Staking reward key \"{}\"", rewards_key_alias);
-    println!(
-        "The ledger node has been setup to use this validator's address and \
-         consensus key."
-    );
-    println!();
-    let genesis_validator = genesis::Validator {
-        pos_data: anoma::ledger::pos::GenesisValidator {
-            address: validator_address,
-            staking_reward_address: rewards_address,
-            tokens: token::Amount::whole(200_000),
-            consensus_key: consensus_key.public.clone(),
-            staking_reward_key: rewards_key.public.clone(),
-        },
-        account_key: validator_key.public.clone(),
-        non_staked_balance: token::Amount::whole(100_000),
-        // TODO replace with https://github.com/anoma/anoma/issues/25)
-        vp_code_path: "wasm/vp_user.wasm".into(),
-    };
-    // TODO print in toml format after we have https://github.com/anoma/anoma/issues/425
-    println!("Genesis validator config: {:#?}", genesis_validator);
 }
