+++ apps/src/lib/client/mod.rs
@@ -1,4 +1,6 @@
+pub mod gossip;
 pub mod rpc;
 pub mod signing;
 mod tendermint_websocket_client;
 pub mod tx;
+pub mod utils;
